+++ CHANGELOG.md
@@ -72,6 +72,8 @@
 * Fixed an issue that the current road name label flashes when camera state changes or travels onto an unnamed road. ([#2958](https://github.com/mapbox/mapbox-navigation-ios/pull/2958))
 * Added the `NavigationOptions.tileStoreConfiguration` property and arguments to `PassiveLocationDataSource(directions:systemLocationManager:tileStoreLocation:)`, `NavigationMapView(frame:navigationCameraType:tileStoreLocation:)`, and `PredictiveCacheManager(predictiveCacheOptions:tileStoreMapOptions:)` for customizing the locations where navigation and map tiles are stored. ([#2956](https://github.com/mapbox/mapbox-navigation-ios/pull/2956))
 * Fixed an issue where lane guidance icons would sometimes highlight the wrong arrow.([#2942](https://github.com/mapbox/mapbox-navigation-ios/pull/2942))
+* Route Duration Annotations feature now supports a list of fonts to be used. This allows for fallback fonts to be specified and more languages to be supported.  ([#2873](https://github.com/mapbox/mapbox-navigation-ios/pull/2873)):
+* Fix an issue where the supplied list of routes was ignored in `showRouteDurations(along:)`.  ([#2873](https://github.com/mapbox/mapbox-navigation-ios/pull/2873))
 
 ## v1.3.0
 
@@ -109,7 +111,7 @@
 * Fixed an issue where the route line periodically peeked out from behind the user puck even though `NavigationViewController.routeLineTracksTraversal` was enabled. ([#2737](https://github.com/mapbox/mapbox-navigation-ios/pull/2737))
 * Created the `UserHaloCourseView` similar to `UserCourseView` for approximate location on iOS 14 during the navigation to represent user location. Allow the switch between `UserHaloCourseView` and `UserCourseView` when precise mode is changed. ([#2664](https://github.com/mapbox/mapbox-navigation-ios/pull/2664))
 * Added option to show route duration callouts when previewing route alternatives ([#2734](https://github.com/mapbox/mapbox-navigation-ios/pull/2734)):
-  * `NavigationMapView.showRouteDurations(along routes:)` to show duration annotation callouts on the map for the provided routes.
+  * `NavigationMapView.showRouteDurations(along:)` to show duration annotation callouts on the map for the provided routes.
   * `NavigationMapView.removeRouteDurations()` to remove any route duration annotations currently displayed on the map.
 
 ### Instruction banners